Increase retries to three
Try reading a DataGroup a total of three times instead of two times. This is needed in case extendedMode is used but fails (first try), then the read size is reduced to 256 bytes, but fails again (second try). Then we need the third try, which sets the read size to 160 bytes.